requirements.txt
numpy
pandas
matplotlib
scikit-learn
certifi
pymongo
python-dotenv
dill
pyaml
# -e .